requirements.txt
flask
flask_jwt_extended
flask_mail
argon2-cffi
flask_pymongo
streamlit
pymongo
mediapipe
opencv-python
scikit-learn
numpy
pandas
streamlit
plotly
flask-socketio
